requirements.txt
pandas
streamlit
jellyfish
scikit-learn
google-generativeai
python-dotenv
thefuzz
python-Levenshtein